Add support for multiple bit widths + Do calculations in 64 bit by default

+++ src/straw/rice/ricer.py
@@ -26,7 +26,7 @@ class Ricer:
     ##########################
 
     def guess_parameter(self, frame_residual: np.array) -> np.int8:
-        smallframe = frame_residual[:self.responsiveness].astype(np.int32)
+        smallframe = frame_residual[:self.responsiveness].astype(np.int64)
         ext_rice.interleave_frame(smallframe)
         param = np.clip(np.log2(smallframe.mean()), 0, (1 << StrawSizes.residual.param) - 1)
         return np.round(param).astype(np.int8)
